requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
python-dotenv==1.0.1

# Dev / testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
